requirements.txt
Django==2.1.5
djangorestframework==3.9.0
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.6.1
pytz==2018.9
